tidied up some of the adminservice

+++ src/main/java/com/revature/repo/AdminRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminRepo extends JpaRepository<Admin, Long> {
+public interface AdminRepo extends JpaRepository<Admin, Integer> {
 
 }
